create setCardHeights()
The MMG card has the most text, so its height property is set to "fit-content," to prevent content from being clipped. I created setCardHeights(), to get the MMG card, or tallestCard, height value, and then set it as the height for the rest of the cards. This function is called on window load and on window resize.