solution for .carousel-dot on click glitch
Okay, so! I realized I was running into issues with the .carousel-dot on click, because obj.imgIndex was about as specific as picIndex was. (Not very.)

In order to target the correct object, I created a .forEach() loop that catalogs dotIndex positions by object. The eighteen dotIndex positions are also assigned to firstDotArr, middleDotArr, and lastDotArr, based on whether they're first, middle, or last dots, in each of the carousels.

So now, an if statement can check whether the clicked dot is considered a first, middle, or last dot, and then decide whether the particular object's imageIndex should be updated to 0, 1, or 2.